Refactor into separate service files
