+++ nite-owl/lib/createPlaylistEvent.js
@@ -1,76 +1,9 @@
-const axios = require("axios").default;
 const util = require("util");
+const nextup = require("./nextup.service");
+const trackSearch = require("./trackSearch.service");
 
-const DURATION_TOLERANCE = 30000;
 const AUTOMATION_USER_ID = "5820844800999424";
 const DJ_PLAY_COOLDOWN_MINUTES = 20;
-
-async function addFreeformPlaylistTrack(data, instance) {
-  const response = await instance.post(`/playlist/freeform`, data, {
-    headers: { "Content-Type": "application/json" },
-  });
-  return response.data;
-}
-
-function getAxiosInstance(apiKey) {
-  const instance = axios.create({
-    baseURL: process.env.API_URL,
-    params: {
-      api_key: apiKey,
-    },
-  });
-  instance.interceptors.response.use(
-    (response) => {
-      return response;
-    },
-    (error) => {
-      console.log(util.inspect(error));
-      console.log(error.response.data);
-      return Promise.reject(error);
-    }
-  );
-  return instance;
-}
-
-async function findTrack(trackTitle, duration, instance) {
-  const durationMs = duration * 1000;
-  const trackLocator = getSearchExecutor(durationMs, instance);
-  const rawCandidates = await trackLocator(trackTitle);
-  if (rawCandidates.length > 0) {
-    return rawCandidates;
-  }
-  let searchTerm = trackTitle;
-  const cleanRegex = /\W{2}clean\W?/gi;
-  const cleanIndex = trackTitle.search(cleanRegex);
-  if (cleanIndex > 0) {
-    searchTerm = trackTitle.slice(0, cleanIndex);
-    const cleanCandidates = await trackLocator(searchTerm);
-    if (cleanCandidates.length > 0) {
-      return cleanCandidates;
-    }
-  }
-  const featuringRegex = /\W[([](featuring|ft|feat).+[)\]]$/gi;
-  const featuringIndex = searchTerm.search(featuringRegex);
-  if (featuringIndex > 0) {
-    searchTerm = searchTerm.slice(0, featuringIndex);
-    const featuringCandidates = await trackLocator(searchTerm);
-    if (featuringCandidates.length > 0) {
-      return featuringCandidates;
-    }
-  }
-  const titleTerms = searchTerm.split(" ");
-  for (let i = 1; i < titleTerms.length; i++) {
-    const offset = 0 - i;
-    const currentTerms = titleTerms.slice(0, offset);
-    const currentSearchTerm = currentTerms.join(" ");
-    const currentCandidates = await trackLocator(currentSearchTerm);
-    if (currentCandidates.length > 0) {
-      return currentCandidates;
-    }
-  }
-
-  return [];
-}
 
 function getXMinutesPrevious(x) {
   const date = new Date();
@@ -78,48 +11,11 @@ function getXMinutesPrevious(x) {
   return date.valueOf();
 }
 
-async function getMostRecentPlays(instance) {
+async function getMostRecentPlays() {
   const params = {
     start: getXMinutesPrevious(DJ_PLAY_COOLDOWN_MINUTES),
   };
-  const response = await instance.get(`/playlist`, {
-    headers: { "Content-Type": "application/json" },
-    params: params,
-  });
-  return response.data;
-}
-
-function getSearchExecutor(duration, instance) {
-  return async (searchTerm) => {
-    console.debug(`Search using term: ${searchTerm}`);
-    let params = {};
-    params.term = searchTerm;
-    params.type = "track";
-    console.log(duration);
-    if (duration !== undefined && duration !== null && !isNaN(duration)) {
-      params["track[duration_ms][gte]"] = duration - DURATION_TOLERANCE;
-      params["track[duration_ms][lte]"] = duration + DURATION_TOLERANCE;
-    }
-    console.debug(`Query params: ${util.inspect(params)}`);
-    const response = await instance.get(`/search`, {
-      headers: { "Content-Type": "application/json" },
-      params: params,
-    });
-    const candidates = response.data.hits.map((hit) => {
-      return {
-        artist:
-          hit._source.track_artist?.name ??
-          hit._source.album.album_artist?.name,
-        album: hit._source.album.title,
-        albumYear: hit._source.album.year,
-        label: hit._source.album.label,
-        currentTags: hit._source.album.album_artist?.current_tags,
-        title: hit._source.title,
-        trackNumber: hit._source.track_num,
-      };
-    });
-    return candidates;
-  };
+  return nextup.getPlaylist(params);
 }
 
 module.exports = async function (req, res) {
@@ -142,9 +38,9 @@ module.exports = async function (req, res) {
   }
 
   console.debug(util.inspect(req.query));
+  nextup.setApiKey(req.query.api_key);
 
-  const instance = getAxiosInstance(req.query.api_key);
-  const recentPlays = await getMostRecentPlays(instance);
+  const recentPlays = await getMostRecentPlays();
   const djPlay = recentPlays.find(
     (playlistEvent) => playlistEvent.selector.id !== AUTOMATION_USER_ID
   );
@@ -154,11 +50,11 @@ module.exports = async function (req, res) {
     return;
   }
 
-  const searchResults = await findTrack(
+  const searchResults = await trackSearch.find(
     req.query.title,
-    req.query.duration,
-    instance
+    req.query.duration
   );
+
   console.debug(`Artist albums: ${util.inspect(searchResults)}`);
   const targetTrack = searchResults.find(
     (result) =>
@@ -185,9 +81,8 @@ module.exports = async function (req, res) {
     categories: targetTrack.currentTags ?? [],
     notes: "Music Mix",
   };
-  const result = await addFreeformPlaylistTrack(
-    JSON.stringify(capturedPlaylistEvent),
-    instance
+  const result = await nextup.addPlaylistEvent(
+    JSON.stringify(capturedPlaylistEvent)
   );
   res.send(result);
 };
